Clean up some vibe code crap
