+++ src/pages/download.page.server.tsx
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { getAllVersions, getLatestVersion } from '../api/api';
 import { PageContextServer } from '../types';
+import { IGithubRelease } from '../types/githubTypes';
 
 async function onBeforeRender(pageContext: PageContextServer) {
     const allVersions = (await getAllVersions()) ?? [];
@@ -8,25 +9,24 @@ async function onBeforeRender(pageContext: PageContextServer) {
         .filter((r) => !r.draft)
         .sort((a, b) => new Date(b.published_at || b.created_at).getTime() - new Date(a.published_at || a.created_at).getTime());
 
-    const latestFromList = ordered[0];
     const data = await getLatestVersion();
-    const latest = data ?? latestFromList;
-    const release = latest ?? null;
 
-    const dmgBuild = release?.assets.find((asset) => asset.name.endsWith('.dmg'));
-    const exeBuild = release?.assets.find((asset) => asset.name.endsWith('.exe'));
-    const debBuild = release?.assets.find((asset) => asset.name.endsWith('.deb'));
-    const rpmBuild = release?.assets.find((asset) => asset.name.endsWith('.rpm'));
+    const dmgBuild = data?.assets.find((asset) => asset.name.endsWith('.dmg'));
+    const exeBuild = data?.assets.find((asset) => asset.name.endsWith('.exe'));
+    const debBuild = data?.assets.find((asset) => asset.name.endsWith('.deb'));
+    const rpmBuild = data?.assets.find((asset) => asset.name.endsWith('.rpm'));
 
-    const zipBuilds = release?.assets.filter((asset) => asset.name.endsWith('.zip'));
+    const zipBuilds = data?.assets.filter((asset) => asset.name.endsWith('.zip'));
     const winZip = zipBuilds?.find((asset) => asset.name.includes('win'));
     const macZip = zipBuilds?.find((asset) => asset.name.includes('mac'));
     const linuxZip = zipBuilds?.find((asset) => asset.name.includes('linux'));
 
-    const previousReleases = ordered.filter((r) => (latest ? r.id !== latest.id : true));
+    const computedChangelog = allVersions.reduce((acc: string, curr: IGithubRelease) => {
+        return `${acc}\n# ${curr.name}\n${curr.body}`;
+    }, '');
 
     // available as `pageContext.pageProps`
-    const pageProps = { dmgBuild, exeBuild, debBuild, rpmBuild, winZip, macZip, linuxZip, latestVersion: latest, previousReleases };
+    const pageProps = { dmgBuild, exeBuild, debBuild, rpmBuild, winZip, macZip, linuxZip, latestVersion: data, computedChangelog };
     return {
         pageContext: {
             pageProps,
