+++ src/pages/nightly.page.tsx
@@ -23,52 +23,10 @@ function Page(pageProps: {
     winZip: IReleaseAsset | undefined;
     macZip: IReleaseAsset | undefined;
     linuxZip: IReleaseAsset | undefined;
-    previousReleases: IGithubRelease[];
+    releaseDate: string | undefined;
+    computedChangelog: string;
 }) {
-    const { latestVersion, dmgBuild, exeBuild, debBuild, rpmBuild, winZip, macZip, linuxZip, previousReleases } = pageProps;
-    const previous = previousReleases ?? [];
-    const formatReleaseDate = (release?: IGithubRelease) => {
-        const dateString = release?.published_at ?? release?.created_at;
-        if (dateString == null) return undefined;
-        const date = new Date(dateString);
-        if (Number.isNaN(date.getTime())) return undefined;
-        return date.toISOString().split('T')[0];
-    };
-    const cleanReleaseBody = (release: IGithubRelease) => {
-        const body = release.body;
-        const dateLabel = formatReleaseDate(release);
-        if (body == null) return '';
-
-        const normalized = body
-            .replace(/\r\n/g, '\n')
-            .replace(/'\s+'/g, '\n')
-            .replace(/'\s*\n/g, '\n')
-            .replace(/\n\s*'/g, '\n');
-
-        const lines = normalized
-            .split('\n')
-            .map((line) => line.trim())
-            .filter((line) => line.length > 0)
-            .filter((line) => !/^Built on\s+\d{4}-\d{2}-\d{2}$/i.test(line))
-            .filter((line) => (dateLabel == null ? true : line !== dateLabel));
-
-        const cleaned = lines
-            .map((line) => {
-                const withoutDate = line.replace(/Built on\s+\d{4}-\d{2}-\d{2}/gi, '').trim();
-                return withoutDate.replace(/\s{2,}/g, ' ').trim();
-            })
-            .filter((line) => line.length > 0);
-
-        if (cleaned.length === 0) return '';
-        return cleaned.map((line) => `- ${line}`).join('\n');
-    };
-
-    const releaseToMarkdown = (release: IGithubRelease) => {
-        const dateHeading = formatReleaseDate(release) ?? release.tag_name ?? 'Nightly build';
-        const body = cleanReleaseBody(release);
-        return body.length > 0 ? `# ${dateHeading}\n\n${body}` : `# ${dateHeading}`;
-    };
-    const nightlyDateLabel = formatReleaseDate(latestVersion);
+    const { latestVersion, dmgBuild, exeBuild, debBuild, rpmBuild, winZip, macZip, linuxZip, releaseDate, computedChangelog } = pageProps;
 
     let currentBuild: IReleaseAsset | undefined;
     let zipBuild: IReleaseAsset | undefined;
@@ -169,9 +127,7 @@ function Page(pageProps: {
                                     }}
                                     fontWeight="bold"
                                 >
-                                    {latestVersion != null
-                                        ? `Download Nightly${nightlyDateLabel != null ? ` (${nightlyDateLabel})` : ''}`
-                                        : 'No nightly release found'}
+                                    {latestVersion != null ? `Download Nightly${releaseDate != null ? ` (${releaseDate})` : ''}` : 'No nightly release found'}
                                 </Text>
                             </HStack>
                             <HStack>
@@ -214,50 +170,12 @@ function Page(pageProps: {
                         w="full"
                         overflowY="scroll"
                     >
-                        {latestVersion != null ? (
-                            <>
-                                <Text
-                                    fontWeight="bold"
-                                    mb={2}
-                                >
-                                    Latest changes
-                                </Text>
-                                <ReactMarkdown
-                                    components={ChakraUIRenderer()}
-                                    skipHtml
-                                >
-                                    {releaseToMarkdown(latestVersion)}
-                                </ReactMarkdown>
-                                {previous.length > 0 && (
-                                    <>
-                                        <Box h={4} />
-                                        <Text
-                                            fontWeight="bold"
-                                            mb={2}
-                                        >
-                                            Previous changes
-                                        </Text>
-                                        <VStack
-                                            align="stretch"
-                                            spacing={6}
-                                        >
-                                            {previous.map((release) => (
-                                                <Box key={release.id}>
-                                                    <ReactMarkdown
-                                                        components={ChakraUIRenderer()}
-                                                        skipHtml
-                                                    >
-                                                        {releaseToMarkdown(release)}
-                                                    </ReactMarkdown>
-                                                </Box>
-                                            ))}
-                                        </VStack>
-                                    </>
-                                )}
-                            </>
-                        ) : (
-                            <Text>No changelog available.</Text>
-                        )}
+                        <ReactMarkdown
+                            components={ChakraUIRenderer()}
+                            skipHtml
+                        >
+                            {computedChangelog}
+                        </ReactMarkdown>
                     </Box>
                 </Box>
             </VStack>
